requirements.txt
<<<<<<< HEAD
<<<<<<< HEAD
black
# flake8
isort
# pip
# python-dotenv
# setuptools
# wheel
click==8.1.3
Markdown==3.4.1
matplotlib==3.6.3
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
pytest==7.2.0
python-dotenv==0.21.0
setuptools==65.6.3
tqdm
-e .
=======
# black
# flake8
# isort
# pip
# python-dotenv
# setuptools
# wheel
=======
black
flake8
isort
pip
python-dotenv
setuptools
wheel
>>>>>>> d5fbdcc (Finish Initial docker)
-e .
>>>>>>> 30a8e1f (Dockerfiles done but I commented all installs because Im flying)
